requirements.txt
streamlit==1.11.0
plotly==5.4.0
numpy==1.21.4
pyarrow==10.0.1
snowflake-snowpark-python
#snowflake-snowpark-python==0.7.0
pandas
